requirements.txt
google-generativeai>=0.8.0
python-dotenv
Pillow<11.0
fastapi
uvicorn[standard]
python-multipart
